change: extract shared scan report functionality
- add common report models and lifecycle helpers
- move shared SQL report loading into report common modules
- rename the scan-specific summary model to scan_report_summary_t
- extract shared GMP report XML serialization
- keep vulnerability result counts and resource serialization scan-specific
- update get_scan_report to use the shared report helpers
