+++ src/manage_scan_report.c
@@ -22,137 +22,18 @@
 #define G_LOG_DOMAIN "md manage"
 
 /**
- * @brief Controls derived from GET data for GET_REPORT.
- */
-typedef struct
-{
-  gchar *zone;
-} get_report_controls_t;
-
-/**
- * @brief Free a GET_REPORT controls structure.
- * @param controls struct to free.
- */
-static void
-get_report_controls_cleanup (get_report_controls_t *controls)
-{
-  if (controls == NULL)
-    return;
-
-  g_free (controls->zone);
-  controls->zone = NULL;
-}
-
-/**
- * @brief Allocate and initialize a report resource summary.
- *
- * @return Newly allocated resource summary.
- */
-static report_resource_summary_t
-report_resource_summary_new (void)
-{
-  return g_malloc0 (sizeof (struct report_resource_summary));
-}
-
-/**
- * @brief Free a report resource summary.
- *
- * @param[in] summary  Resource summary to free.
- */
-static void
-report_resource_summary_free (report_resource_summary_t summary)
-{
-  g_free (summary);
-}
-
-/**
- * @brief Allocate and initialize a report target reference.
- *
- * @return Newly allocated target reference.
- */
-static report_target_reference_t
-report_target_reference_new (void)
-{
-  report_target_reference_t target;
-
-  target = g_malloc0 (sizeof (struct report_target_reference));
-  target->type = REPORT_TARGET_TYPE_NONE;
-
-  return target;
-}
-
-/**
- * @brief Free a report target reference.
- *
- * @param[in] target  Target reference to free.
- */
-static void
-report_target_reference_free (report_target_reference_t target)
-{
-  if (target == NULL)
-    return;
-
-  g_free (target->uuid);
-  g_free (target->name);
-  g_free (target->comment);
-
-  g_free (target);
-}
-
-/**
- * @brief Allocate and initialize a report task reference.
- *
- * @return Newly allocated task reference.
- */
-static report_task_reference_t
-report_task_reference_new (void)
-{
-  report_task_reference_t task;
-
-  task = g_malloc0 (sizeof (struct report_task_reference));
-
-  task->target = report_target_reference_new ();
-  if (task->target == NULL)
-    {
-      g_free (task);
-      return NULL;
-    }
-
-  return task;
-}
-
-/**
- * @brief Free a report task reference.
- *
- * @param[in] task  Task reference to free.
- */
-static void
-report_task_reference_free (report_task_reference_t task)
-{
-  if (task == NULL)
-    return;
-
-  g_free (task->uuid);
-  g_free (task->name);
-  g_free (task->comment);
-  g_free (task->usage_type);
-
-  report_target_reference_free (task->target);
-
-  g_free (task);
-}
-
-/**
  * @brief Allocate and initialize a report summary.
  *
  * @return Newly allocated report summary, or NULL on allocation failure.
  */
-report_summary_t
-report_summary_new (void)
+scan_report_summary_t
+scan_report_summary_new (void)
 {
-  report_summary_t summary;
+  scan_report_summary_t summary;
 
-  summary = g_malloc0 (sizeof (struct report_summary));
+  summary = g_malloc0 (sizeof (struct scan_report_summary));
+
+  summary->base = report_summary_base_new ();
 
   summary->task = report_task_reference_new ();
   if (summary->task == NULL)
@@ -178,23 +59,12 @@ report_summary_new (void)
  * @param[in] summary  Report summary to free.
  */
 void
-report_summary_free (report_summary_t summary)
+scan_report_summary_free (scan_report_summary_t summary)
 {
   if (summary == NULL)
     return;
 
-  g_free (summary->id);
-  g_free (summary->name);
-  g_free (summary->comment);
-  g_free (summary->owner_name);
-  g_free (summary->timestamp);
-  g_free (summary->scan_start);
-  g_free (summary->scan_end);
-
-  g_free (summary->scan_run_status_str);
-  g_free (summary->timezone);
-  g_free (summary->timezone_abbrev);
-
+  report_summary_base_free (summary->base);
   report_task_reference_free (summary->task);
   report_resource_summary_free (summary->resources);
 
@@ -322,10 +192,10 @@ resolve_get_report_controls (const get_data_t *get,
  */
 manage_get_scan_report_response_t
 manage_get_scan_report_summary (const get_data_t *get,
-                                report_summary_t *summary)
+                                scan_report_summary_t *summary)
 {
   get_report_controls_t controls = {0};
-  report_summary_t loaded_model = NULL;
+  scan_report_summary_t loaded_model = NULL;
   report_t report = 0;
   int ret;
 
@@ -365,7 +235,7 @@ manage_get_scan_report_summary (const get_data_t *get,
       return MANAGE_GET_SCAN_REPORT_ERROR;
     }
 
-  loaded_model = report_summary_new ();
+  loaded_model = scan_report_summary_new ();
   if (loaded_model == NULL)
     {
       get_report_controls_cleanup (&controls);
@@ -381,46 +251,11 @@ manage_get_scan_report_summary (const get_data_t *get,
 
   if (ret)
     {
-      report_summary_free (loaded_model);
+      scan_report_summary_free (loaded_model);
       return MANAGE_GET_SCAN_REPORT_ERROR;
     }
 
   *summary = loaded_model;
 
   return MANAGE_GET_SCAN_REPORT_SUCCESS;
-}
-
-/**
- * @brief Get the string representation of a report target type.
- *
- * @param[in] type  Report target type.
- *
- * @return String representation of the report target type.
- */
-const gchar *
-report_target_type_to_string (report_target_type_t type)
-{
-  switch (type)
-    {
-    case REPORT_TARGET_TYPE_NONE:
-      return "none";
-
-    case REPORT_TARGET_TYPE_TARGET:
-      return "target";
-
-    case REPORT_TARGET_TYPE_OCI_IMAGE:
-      return "oci_image";
-
-    case REPORT_TARGET_TYPE_WEB_APPLICATION:
-      return "web_application";
-
-    case REPORT_TARGET_TYPE_AGENT_GROUP:
-      return "agent_group";
-
-    case REPORT_TARGET_TYPE_IMPORT:
-      return "import";
-
-    default:
-      return "unknown";
-    }
 }
