+++ src/manage_scan_report.h
@@ -12,24 +12,7 @@
 #define _GVMD_MANAGE_SCAN_REPORT_H
 
 #include "manage.h"
-
-/**
- * @brief Full and filtered values for a report count.
- */
-typedef struct
-{
- int full;
- int filtered;
-} report_count_t;
-
-/**
- * @brief Full and filtered report severity.
- */
-typedef struct
-{
- double full;
- double filtered;
-} report_severity_t;
+#include "manage_report_common.h"
 
 /**
  * @brief Result summary for a normal vulnerability report.
@@ -54,71 +37,6 @@ typedef struct
 } report_result_summary_t;
 
 /**
- * @brief Resource counts belonging to a report.
- */
-struct report_resource_summary
-{
- int hosts;
- int ports;
- int applications;
- int operating_systems;
- int vulnerabilities;
- int cves;
- int closed_cves;
- int tls_certificates;
- int errors;
-};
-
-typedef struct report_resource_summary* report_resource_summary_t;
-
-/**
- * @brief Type of resource scanned by the report task.
- */
-typedef enum
-{
- REPORT_TARGET_TYPE_NONE = 0,
- REPORT_TARGET_TYPE_TARGET,
- REPORT_TARGET_TYPE_OCI_IMAGE,
- REPORT_TARGET_TYPE_WEB_APPLICATION,
- REPORT_TARGET_TYPE_AGENT_GROUP,
- REPORT_TARGET_TYPE_IMPORT
-} report_target_type_t;
-
-/**
- * @brief Reference to the resource scanned by a report task.
- */
-struct report_target_reference
-{
- report_target_type_t type;
-
- resource_t id;
- gchar* uuid;
- gchar* name;
- gchar* comment;
-
- int in_trash;
-};
-
-typedef struct report_target_reference* report_target_reference_t;
-
-/**
- * @brief Reference to the task associated with a report.
- */
-struct report_task_reference
-{
- task_t id;
- gchar* uuid;
- gchar* name;
- gchar* comment;
- gchar* usage_type;
- int progress;
-
- report_target_reference_t target;
-};
-
-typedef struct report_task_reference* report_task_reference_t;
-
-/**
  * @brief Structured model of a normal report.
  *
  * This model intentionally excludes:
@@ -130,35 +48,18 @@ typedef struct report_task_reference* report_task_reference_t;
  * - report export and report format data
  * - temporary report-rendering state
  */
-struct report_summary
+struct scan_report_summary
 {
- report_t report;
- user_t owner;
-
- gchar* id;
- gchar* name;
- gchar* comment;
- gchar* owner_name;
-
- time_t creation_time;
- time_t modification_time;
-
- gchar* timestamp;
- gchar* scan_start;
- gchar* scan_end;
- task_status_t scan_run_status;
- gchar* scan_run_status_str;
-
- gchar* timezone;
- gchar* timezone_abbrev;
+ report_summary_base_t base;
 
  report_task_reference_t task;
 
  report_resource_summary_t resources;
+
  report_result_summary_t results;
 };
 
-typedef struct report_summary* report_summary_t;
+typedef struct scan_report_summary* scan_report_summary_t;
 
 /**
  * @brief Result of loading a structured report model.
@@ -172,17 +73,14 @@ typedef enum
  MANAGE_GET_SCAN_REPORT_UNSUPPORTED_TYPE = 3
 } manage_get_scan_report_response_t;
 
-report_summary_t
-report_summary_new (void);
+scan_report_summary_t
+scan_report_summary_new (void);
 
 void
-report_summary_free (report_summary_t);
+scan_report_summary_free (scan_report_summary_t);
 
 manage_get_scan_report_response_t
 manage_get_scan_report_summary (const get_data_t *,
-                                report_summary_t *);
-
-const gchar *
-report_target_type_to_string (report_target_type_t);
+                                scan_report_summary_t *);
 
 #endif /* _GVMD_MANAGE_SCAN_REPORT_H */
