+++ src/manage_sql_scan_report.h
@@ -17,17 +17,6 @@ int
 manage_sql_fill_report_summary (report_t,
                                 const get_data_t *,
                                 const gchar *,
-                                report_summary_t);
-int
-report_scan_run_status (report_t, task_status_t*);
-
-int
-report_vuln_count (report_t);
-
-char*
-scan_end_time (report_t report);
-
-char*
-scan_start_time (report_t report);
+                                scan_report_summary_t);
 
 #endif /* _GVMD_MANAGE_SQL_SCAN_REPORT_H */
